# Conflicts:
#	GamePlay.cpp

+++ CartelPosicion.cpp
@@ -2,12 +2,12 @@
 #include "CartelPosicion.h"
 
 
-void CartelPosicion::showPositionOnScreen(std::string cartel, sf::Vector2f posicionCartel, sf::Vector2f posicionElemento)
+void CartelPosicion::showPositionOnScreen(std::string cartel, sf::Vector2f  posicionCartel, sf::Vector2f  posicionElemento)
 {
   fuente.loadFromFile("assets/fonts/monofonto_rg.otf");
   _cartel.setFont(fuente);
   _cartel.setPosition(posicionCartel);
-  _cartel.setString(cartel + "\nPosicion: \nx = " +  std::to_string(posicionElemento.x) +  "\ny = " + std::to_string(posicionElemento.y));
+  _cartel.setString(cartel + "\nPosicion: \nx = " +  std::to_string((int)posicionElemento.x) +  "\ny = " + std::to_string((int)posicionElemento.y));
   _cartel.setCharacterSize(10);
 }
